Move ranking system plugins to frontend app
