+++ packages/engine/src/index.ts
@@ -3,4 +3,4 @@ export * from "./types/poll";
 export * from "./types/channel";
 export * from "./questionTypeStore";
 export * from "./themeStore";
-export * from "./ranking-systems";
+export * from "./rankingSystemStore";
