--- packages/frontend/app/plugins/index.ts
@@ -1,1 +1,0 @@
-import "./questionTypes";
