+++ packages/frontend/app/plugins/rankingSystems/index.ts
@@ -1,0 +1,6 @@
+import { rankingSystemStore } from "@marcianosrs/engine";
+import { pointRankingSystem } from "./pointRankingSystem";
+
+export const registerRankingPlugins = () => {
+	rankingSystemStore.add(pointRankingSystem);
+};
