+++ packages/frontend/app/routes/_index.tsx
@@ -14,7 +14,6 @@ import {
 } from "~/util/session.server";
 import { getChannels } from "./api.server";
 import type { ChannelDTO } from "@marcianosrs/engine";
-import "../plugins";
 
 export const meta: MetaFunction = () => {
 	return [{ title: "New Remix App" }];
