+++ packages/utils/src/pluginStore.ts
@@ -1,14 +1,20 @@
 export const createPluginStore = <BasePlugin, AddType = BasePlugin>(
-    getIdentifier: (plugin: BasePlugin) => string
+	getIdentifier: (plugin: BasePlugin) => string
 ) => {
-    const plugins: BasePlugin[] = [];
+	const plugins: BasePlugin[] = [];
 
-    return {
-        add: <Plugin extends AddType>(type: Plugin) => {
-            plugins.push(type as unknown as BasePlugin);
-        },
-        get: (pluginId: string): BasePlugin | undefined =>
-            plugins.find((e) => getIdentifier(e) === pluginId),
-        getIdentifiers: () => plugins.map(getIdentifier),
-    };
+	return {
+		add: <Plugin extends AddType>(type: Plugin) => {
+			const identifier = getIdentifier(type);
+			if (
+				plugins.some((plugin) => getIdentifier(plugin) === identifier)
+			) {
+				return;
+			}
+			plugins.push(type as unknown as BasePlugin);
+		},
+		get: (pluginId: string): BasePlugin | undefined =>
+			plugins.find((e) => getIdentifier(e) === pluginId),
+		getIdentifiers: () => plugins.map(getIdentifier),
+	};
 };
